requirements.txt
streamlit>=1.32.0
streamlit-image-coordinates>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
